requirements.txt
gradio>=4.44.1
huggingface_hub>=0.23.0
gtts>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
